+++ R/add_coding_scheme.R
@@ -11,6 +11,10 @@
 #' @export
 add_coding_scheme <- function(units, filter_has_codes = TRUE, overwrite = FALSE) {
   cli_setting()
+  # input validation
+  checkmate::assert_tibble(units)
+  checkmate::assert_logical(filter_has_codes, len = 1)
+  checkmate::assert_logical(overwrite, len = 1)
 
   # Conserve attributes
   unit_attributes <- attributes(units)
@@ -21,6 +25,9 @@ add_coding_scheme <- function(units, filter_has_codes = TRUE, overwrite = FALSE)
 
     return(units)
   }
+
+  units_cols <- c("unit_key", "ws_id", "unit_id", "coding_scheme", "unit_variables")
+  assert_cols(units, units_cols, "units")
 
   unit_keys <- units$unit_key