+++ R/login_studio.R
@@ -32,6 +32,12 @@ login_studio <- function(base_url = "https://www.iqb-studio.de/",
                          dialog = TRUE,
                          verbose = FALSE) {
   cli_setting()
+  checkmate::assert_character(base_url, len = 1)
+  checkmate::assert_character(app_version, len = 1)
+  checkmate::assert_logical(keyring, len = 1)
+  checkmate::assert_logical(change_key, len = 1)
+  checkmate::assert_logical(dialog, len = 1)
+  checkmate::assert_logical(verbose, len = 1)
 
   # Authentication
   credentials <- get_credentials(base_url = base_url,